feat: Enhance data accuracy in models and controllers
Implemented various updates across models and controllers to ensure the accuracy of data handling and management:

- Updated the product model to enforce data integrity and receive accurate product information.
- Revised the invoice model to incorporate a field for the total value of customer discounts, providing comprehensive invoice details.
- Enhanced the customer type model to ensure accurate customer type data management.
- Made significant updates to the sales (invoice) controllers, including create, edit, and read single invoice controllers, to facilitate accurate invoice creation, editing, and retrieval. These updates ensure alignment with frontend requirements for seamless interaction.
- Implemented minor updates in the inventory controllers to improve data consistency and handling.

These changes enhance the reliability and accuracy of data processing across the application, ensuring optimal performance and user experience.

+++ backend/models/inventory/products.js
@@ -36,35 +36,35 @@ const Products = sequelize.define('products', {
         allowNull: true,
     },
     unit_value:{
-        type: DataTypes.DECIMAL,
+        type: DataTypes.FLOAT,
         allowNull: true,
     },
     cost_unit: {
-        type: DataTypes.DECIMAL,
+        type: DataTypes.FLOAT,
         allowNull: false,
     },
     retail_price_unit: {
-        type: DataTypes.DECIMAL,
+        type: DataTypes.FLOAT,
         allowNull: false,
     },
     wholesale_price_unit: {
-        type: DataTypes.DECIMAL,
+        type: DataTypes.FLOAT,
         allowNull: false,
     },
     pieces_per_unit:{
-        type: DataTypes.DECIMAL,
+        type: DataTypes.FLOAT,
         allowNull: false,
     },
     cost_piece:{
-        type: DataTypes.DECIMAL,
+        type: DataTypes.FLOAT,
         allowNull: true,
     },
     retail_price_piece:{
-        type: DataTypes.DECIMAL,
+        type: DataTypes.FLOAT,
         allowNull: true,
     },
     wholesale_price_piece:{
-        type: DataTypes.DECIMAL,
+        type: DataTypes.FLOAT,
         allowNull: true,
     },
     qty: {
@@ -72,10 +72,10 @@ const Products = sequelize.define('products', {
         allowNull: false,
     },
     sale_price_unit: {
-        type: DataTypes.DECIMAL,
+        type: DataTypes.FLOAT,
     },
     sale_price_piece: {
-        type: DataTypes.DECIMAL,
+        type: DataTypes.FLOAT,
     },
     on_sale: {
         type: DataTypes.BOOLEAN,
